Move risks to OSCAL types
