+++ src/components/risk/RiskCharacterizationsTab.vue
@@ -202,8 +202,7 @@
 import { computed, ref } from 'vue';
 import Dialog from '@/volt/Dialog.vue';
 import { useToast } from 'primevue/usetoast';
-import type { Risk } from '@/stores/plan-of-action-and-milestones';
-import type { Characterization, Facet } from '@/oscal';
+import type { Characterization, Facet, Risk } from '@/oscal';
 import { cloneFnJSON as cloneDeep } from '@vueuse/core';
 
 const DEFAULT_FACET_SYSTEM = 'custom';
