+++ src/components/risk/RiskMitigationsTab.vue
@@ -189,8 +189,7 @@
 import { computed, reactive, ref } from 'vue';
 import Dialog from '@/volt/Dialog.vue';
 import { useToast } from 'primevue/usetoast';
-import type { Risk } from '@/stores/plan-of-action-and-milestones';
-import type { MitigatingFactor, SubjectReference } from '@/oscal';
+import type { MitigatingFactor, SubjectReference, Risk } from '@/oscal';
 import { cloneFnJSON as cloneDeep } from '@vueuse/core';
 
 const DEFAULT_SUBJECT_TYPE = 'subject';
